Model changes for AssignmentRoster
